backend/requirements.txt
fastapi[standard]
python-dotenv
openai
pandas
requests
matplotlib